+++ pkg/kubelet/util.go
@@ -32,7 +32,7 @@ func CapacityFromMachineInfo(info *cadvisorApi.MachineInfo) api.ResourceList {
 			int64(info.NumCores*1000),
 			resource.DecimalSI),
 		api.ResourceMemory: *resource.NewQuantity(
-			info.MemoryCapacity,
+			int64(info.MemoryCapacity),
 			resource.BinarySI),
 	}
 	return c
